authorzization and limit request

+++ src/main/java/com/backend_app_hit/app_hit/controller/PostController.java
@@ -1,12 +1,12 @@
 package com.backend_app_hit.app_hit.controller;
 
+import java.time.Duration;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
 import com.backend_app_hit.app_hit.dao.Post;
 import com.backend_app_hit.app_hit.dao.User;
-import com.backend_app_hit.app_hit.exception.InvalidException;
 import com.backend_app_hit.app_hit.exception.NotFoundException;
 import com.backend_app_hit.app_hit.models.PostResponse;
 import com.backend_app_hit.app_hit.repository.PostRepository;
@@ -26,9 +26,16 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import io.github.bucket4j.Bandwidth;
+import io.github.bucket4j.Bucket;
+import io.github.bucket4j.Bucket4j;
+import io.github.bucket4j.Refill;
+
 @RestController
 @RequestMapping("/api/v1/post")
 public class PostController {
+
+    private final Bucket bucket;
 
     @Autowired
     private PostRepository postRepository;
@@ -36,9 +43,14 @@ public class PostController {
     @Autowired
     private UserRepository userRepository;
 
+    public PostController() {
+        Bandwidth limit = Bandwidth.classic(10, Refill.greedy(10, Duration.ofMinutes(1)));
+        this.bucket = Bucket4j.builder().addLimit(limit).build();
+    }
+
     @PostMapping("/create")
     public ResponseEntity<?> createPost(@RequestBody String content) {
-        try {
+        if (bucket.tryConsume(1)) {
             String username = GetUserNameByContext.getUserName();
 
             Post post = new Post();
@@ -56,70 +68,72 @@ public class PostController {
 
             List<Post> posts = new ArrayList<Post>();
             posts.add(post);
-            return ResponseEntity.status(HttpStatus.CREATED).body(new PostResponse(201, "Thêm thành công", username, posts));
-        } catch (Exception e) {
-            throw new InvalidException(e.getMessage());
+            return ResponseEntity.status(HttpStatus.CREATED)
+                    .body(new PostResponse(201, "Thêm thành công", username, posts));
         }
+        return ResponseEntity.status(HttpStatus.TOO_MANY_REQUESTS).body("TOO_MANY_REQUESTS");
 
     }
 
     @GetMapping("/me")
     public ResponseEntity<?> postMe() {
+        String username = GetUserNameByContext.getUserName();
 
-        try {
-            String username = GetUserNameByContext.getUserName();
-
-            Optional<User> uOptional = userRepository.findByUsername(username);
-            if (!uOptional.isPresent()) {
-                throw new UsernameNotFoundException("Username không tồn tại");
-            }
-
-            User user = uOptional.get();
-
-            List<Post> posList = postRepository.findByUserId(user.getId());
-
-            return ResponseEntity.ok(new PostResponse(200,"Thành công",user.getUsername(), posList));
-        } catch (Exception e) {
-            throw new InvalidException(e.getMessage());
+        Optional<User> uOptional = userRepository.findByUsername(username);
+        if (!uOptional.isPresent()) {
+            throw new UsernameNotFoundException("Username không tồn tại");
         }
+
+        User user = uOptional.get();
+
+        List<Post> posList = postRepository.findByUserId(user.getId());
+
+        return ResponseEntity.ok(new PostResponse(200, "Thành công", user.getUsername(), posList));
     }
 
     @PatchMapping(value = "/{postId}")
     public ResponseEntity<?> updatePost(@RequestBody String content, @PathVariable Long postId) {
-        String userName = GetUserNameByContext.getUserName();
-        Optional<Post> postOptional = postRepository.findById(postId);
 
-        if (!postOptional.isPresent()) {
-            throw new NotFoundException("Bài viết không tồn tại");
+        if (bucket.tryConsume(1)) {
+            String userName = GetUserNameByContext.getUserName();
+            Optional<Post> postOptional = postRepository.findById(postId);
+
+            if (!postOptional.isPresent()) {
+                throw new NotFoundException("Bài viết không tồn tại");
+            }
+
+            Post post = postOptional.get();
+
+            if (!post.getUser().getUsername().equals(userName)) {
+                throw new NotFoundException("Không có quyền sửa bài viết này");
+            }
+
+            post.setContent(content);
+            postRepository.save(post);
+            return ResponseEntity.status(HttpStatus.OK).body(post);
         }
-
-        Post post = postOptional.get();
-
-        if (!post.getUser().getUsername().equals(userName)) {
-            throw new NotFoundException("Không có quyền sửa bài viết này");
-        }
-
-        post.setContent(content);
-        postRepository.save(post);
-        return ResponseEntity.status(HttpStatus.OK).body(post);
+        return ResponseEntity.status(HttpStatus.TOO_MANY_REQUESTS).body("TOO_MANY_REQUESTS");
 
     }
 
     @DeleteMapping("/{postId}")
-    public ResponseEntity<?> deletePost(@PathVariable Long postId) {
-        String userName = GetUserNameByContext.getUserName();
-        Optional<Post> postOptional = postRepository.findById(postId);
+    public ResponseEntity<?> deletePost(@PathVariable("postId") Long postId) {
+        if (bucket.tryConsume(1)) {
+            String userName = GetUserNameByContext.getUserName();
+            Optional<Post> postOptional = postRepository.findById(postId);
 
-        if (!postOptional.isPresent()) {
-            throw new NotFoundException("Bài viết không tồn tại");
+            if (!postOptional.isPresent()) {
+                throw new NotFoundException("Bài viết không tồn tại");
+            }
+
+            Post post = postOptional.get();
+
+            if (!post.getUser().getUsername().equals(userName)) {
+                throw new NotFoundException("Không có quyền xoá bài viết này");
+            }
+            postRepository.deleteById(postId);
+            return ResponseEntity.status(HttpStatus.OK).body("Xoá thành công");
         }
-
-        Post post = postOptional.get();
-
-        if (!post.getUser().getUsername().equals(userName)) {
-            throw new NotFoundException("Không có quyền xoá bài viết này");
-        }
-        postRepository.deleteById(postId);
-        return ResponseEntity.status(HttpStatus.OK).body("Xoá thành công");
+        return ResponseEntity.status(HttpStatus.TOO_MANY_REQUESTS).body("TOO_MANY_REQUESTS");
     }
 }
